% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimbWith-package.R
\docType{package}
\name{ClimbWith-package}
\alias{ClimbWith}
\alias{ClimbWith-package}
\title{ClimbWith: Find Rock Climbing Gyms}
\description{
Look for rock climbing gyms based on features you want
}
\author{
\strong{Maintainer}: Gus Lipkin \email{gus@climbwith.info}

}
\keyword{internal}
